requirements.txt
asgiref==3.3.1
Babel==2.9.0
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-ckeditor==6.0.0
django-countries==7.0
django-crispy-forms==1.11.2
django-js-asset==1.2.2
django-phonenumber-field==5.0.0
idna==2.10
phonenumbers==8.12.20
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
stripe==2.56.0
urllib3==1.26.4